+++ fiftyone/core/frames.py
@@ -17,12 +17,12 @@ class NoDatasetFrames(defaultdict):
         super().__init__(FrameSample)
 
     def __getitem__(self, key):
-        if not isinstance(key, six.integer_types):
+        if not isinstance(key, six.integer_types) or key < 0:
             raise FrameError("invalid frame key")
         return super().__getitem__(key)
 
     def __setitem__(self, key, value):
-        if not isinstance(key, six.integer_types):
+        if not isinstance(key, six.integer_types) or key < 0:
             raise FrameError("invalid frame key")
 
         if not isinstance(value, FrameSample):
