+++ fiftyone/core/odm/__init__.py
@@ -18,4 +18,5 @@ from .sample import (
     SampleDocument,
     DatasetSampleDocument,
     NoDatasetSampleDocument,
+    get_implied_field_kwargs,
 )
